--- sites/org/tailwind.config.js
@@ -1,4 +1,0 @@
-// Can't seem to make this work as ESM
-const config = require('../shared/config/tailwind.config.js')
-
-module.exports = config
