+++ sites/org/tailwind.config.mjs
@@ -1,0 +1,3 @@
+import config from '../shared/config/tailwind.config.mjs'
+
+export default config
