+++ sites/shared/components/wordmark.mjs
@@ -1,4 +1,5 @@
 import Link from 'next/link'
+import { useTheme } from 'shared/hooks/use-theme.mjs'
 
 export const colors = [
   'red',
@@ -13,20 +14,24 @@ export const colors = [
   'purple',
 ]
 
-export const InnerWordMark = () => (
-  <span style={{ letterSpacing: '-0.1rem' }}>
-    <span className="text-red-400 hover:text-purple-400">F</span>
-    <span className="text-orange-400 hover:text-violet-400">r</span>
-    <span className="text-yellow-400 hover:text-indigo-400">e</span>
-    <span className="text-lime-400 hover:text-blue-400">e</span>
-    <span className="text-green-400 hover:text-cyan-400">S</span>
-    <span className="text-cyan-400 hover:text-green-400">e</span>
-    <span className="text-blue-400 hover:text-lime-400">w</span>
-    <span className="text-indigo-400 hover:text-yellow-400">i</span>
-    <span className="text-violet-400 hover:text-orange-400">n</span>
-    <span className="text-purple-400 hover:text-red-400">g</span>
-  </span>
-)
+export const InnerWordMark = () => {
+  const { spectrum } = useTheme()
+
+  return (
+    <span style={{ letterSpacing: '-0.1rem' }}>
+      <span className={`text-${spectrum[0]} hover:text-${spectrum[9]}`}>F</span>
+      <span className={`text-${spectrum[1]} hover:text-${spectrum[8]}`}>r</span>
+      <span className={`text-${spectrum[2]} hover:text-${spectrum[7]}`}>e</span>
+      <span className={`text-${spectrum[3]} hover:text-${spectrum[6]}`}>e</span>
+      <span className={`text-${spectrum[4]} hover:text-${spectrum[5]}`}>S</span>
+      <span className={`text-${spectrum[5]} hover:text-${spectrum[4]}`}>e</span>
+      <span className={`text-${spectrum[6]} hover:text-${spectrum[3]}`}>w</span>
+      <span className={`text-${spectrum[7]} hover:text-${spectrum[2]}`}>i</span>
+      <span className={`text-${spectrum[8]} hover:text-${spectrum[1]}`}>n</span>
+      <span className={`text-${spectrum[9]} hover:text-${spectrum[0]}`}>g</span>
+    </span>
+  )
+}
 
 export const WordMark = () => (
   <Link
