+++ sites/shared/components/wrappers/page.mjs
@@ -36,7 +36,7 @@ export const PageWrapper = (props) => {
    * This forces a re-render upon initial bootstrap of the app
    * This is needed to avoid hydration errors because theme can't be set reliably in SSR
    */
-  const [theme] = useTheme()
+  const { theme } = useTheme()
   const [currentTheme, setCurrentTheme] = useState()
   useEffect(() => setCurrentTheme(theme), [currentTheme, theme])
