+++ sites/shared/config/tailwind-force.html
@@ -105,3 +105,29 @@
 
 <!-- toc list padding -->
 <li className="pl-3"></li>
+
+<!-- light theme -->
+<div className='text-violet-300 hover:bg-violet-300 bg-violet-300'></div>
+<!-- dark theme -->
+<div className='text-violet-400 hover:bg-violet-400'></div>
+<!-- aqua theme -->
+<div className='text-blue-200 hover:bg-blue-200 bg-blue-200'></div>
+<svg class="text-blue-600"/>
+<svg class="text-blue-700"/>
+<!-- lgbtq theme -->
+<svg class="text-pink-500"/>
+<!-- monochrome theme -->
+<div className='text-stone-300 hover:bg-stone-300 bg-stone-300'></div>
+<svg class="text-stone-300"/>
+<svg class="text-stone-400"/>
+<svg class="text-stone-500"/>
+<svg class="text-stone-600"/>
+<svg class="text-stone-700"/>
+<svg class="text-stone-800"/>
+<!-- pastel theme -->
+<div className='text-indigo-300 hover:bg-indigo-300 bg-indigo-300'></div>
+<!-- hax0r theme -->
+<div className='bg-lime-500 text-lime-500 hover:bg-lime-500'></div>
+
+
+
