+++ sites/shared/hooks/use-theme.mjs
@@ -1,3 +1,4 @@
+import { spectrum, rating } from 'shared/themes/index.mjs'
 import createPersistedState from 'use-persisted-state'
 
 const usePersistedTheme = createPersistedState('fs-theme')
@@ -11,4 +12,13 @@ const preferredTheme = () => {
   return prefersDarkMode ? 'dark' : 'light'
 }
 
-export const useTheme = () => usePersistedTheme(preferredTheme)
+export const useTheme = () => {
+  const theme = usePersistedTheme(preferredTheme)
+
+  return {
+    theme: theme[0],
+    setTheme: theme[1],
+    spectrum: spectrum[theme[0]],
+    rating: rating[theme[0]],
+  }
+}
