+++ sites/shared/i18n/themes/en.yaml
@@ -5,4 +5,5 @@ lgbtqTheme: LGBTQ Theme
 transTheme: Trans Theme
 pastelTheme: Pastel Theme
 aquaTheme: Aqua Theme
+monochromeTheme: Monochrome Theme
 chooseYourTheme: Choose your theme
