+++ sites/shared/themes/dark.mjs
@@ -1,6 +1,10 @@
-const colors = require('tailwindcss/colors')
+import colors from 'tailwindcss/colors'
 
-module.exports = {
+export const spectrum = [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11].map((i) => 'violet-400')
+
+export const rating = ['green-400', 'yellow-400', 'amber-500', 'orange-500', 'red-500']
+
+export const theme = {
   fontFamily:
     '-apple-system, BlinkMacSystemFont, Segoe UI, Roboto, Oxygen, Ubuntu, Cantarell, Fira Sans, Droid Sans, Helvetica Neue, sans-serif',
