+++ sites/shared/themes/hax0r.mjs
@@ -1,48 +1,47 @@
-const colors = require('tailwindcss/colors')
+import colors from 'tailwindcss/colors'
+
+/*
+ * See the light theme for an example with inline comments
+ */
 
 const bg = '#002808'
 
-module.exports = {
-  fontFamily: `ui-monospace, Menlo, Monaco, "Cascadia Mono", "Segoe UI Mono", "Roboto Mono", "Oxygen Mono", "Ubuntu Monospace", "Source Code Pro", "Fira Mono", "Droid Sans Mono", "Courier New", monospace;`,
+export const spectrum = [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11].map((i) => 'lime-500')
 
+export const rating = [0, 1, 2, 3, 4].map((i) => 'lime-500')
+
+export const theme = {
+  fontFamily: `ui-monospace, Menlo, Monaco, "Cascadia Mono", "Segoe UI Mono", "Roboto Mono", "Oxygen Mono", "Ubuntu Monospace", "Source Code Pro", "Fira Mono", "Droid Sans Mono", "Courier New", monospace;`,
   'base-100': bg,
   'base-200': colors.lime['900'],
   'base-300': colors.lime['800'],
   'base-content': colors.lime['500'],
-
   primary: colors.lime['700'],
   'primary-focus': colors.lime['600'],
   'primary-content': colors.lime['50'],
-
   secondary: colors.lime['600'],
   'secondary-focus': colors.lime['500'],
   'secondary-content': bg,
-
   accent: colors.lime['800'],
   'accent-focus': colors.lime['600'],
   'accent-content': colors.lime['400'],
-
   neutral: bg,
   'neutral-focus': colors.lime['600'],
   'neutral-content': colors.lime['200'],
-
   info: colors.lime['700'],
   success: colors.lime['700'],
   warning: colors.lime['700'],
   error: colors.lime['700'],
-
   '--btn-info-content': colors.teal[300],
   '--btn-success-content': colors.green[300],
   '--btn-warning-content': colors.orange[300],
   '--btn-error-content': colors.red[300],
-
   '--theme-gradient': `repeating-linear-gradient(
     90deg,
     ${colors.lime['900']},
     ${colors.lime['500']} 50%,
     ${colors.lime['900']} 100%
   )`,
-
   '--code-background-color': '#002407',
   '--code-background-highlight-color': '#191919',
   '--code-border-color': colors.lime['900'],
@@ -65,7 +64,6 @@ module.exports = {
   '--code-color-tag': colors.lime['400'],
   '--code-color-property': colors.lime['200'],
   '--code-font-weight-property': 'bold',
-
   '--rounded-box': '0',
   '--rounded-btn': '0',
   '--rounded-badge': '0',
@@ -77,7 +75,6 @@ module.exports = {
   '--border-btn': '1px',
   '--focus-ring': '2px',
   '--focus-ring-offset': '2px',
-
   '--pattern-bg': colors.lime['900'],
   '--pattern-fabric': colors.neutral['400'],
   '--pattern-lining': colors.emerald['500'],
@@ -87,7 +84,6 @@ module.exports = {
   '--pattern-mark': colors.blue['500'],
   '--pattern-contrast': colors.pink['500'],
   '--pattern-note': colors.violet['500'],
-
   '--pattern-dev-0': colors.red['500'],
   '--pattern-dev-1': colors.green['500'],
   '--pattern-dev-2': colors.blue['500'],
@@ -96,7 +92,6 @@ module.exports = {
   '--pattern-dev-5': colors.violet['500'],
   '--pattern-dev-6': colors.teal['500'],
   '--pattern-dev-7': colors.neutral['500'],
-
   '--pattern-text-xs': '0.2rem',
   '--pattern-text-sm': '0.3rem',
   '--pattern-text': '0.4rem',
@@ -105,7 +100,6 @@ module.exports = {
   '--pattern-text-2xl': '1.5rem',
   '--pattern-text-3xl': '2rem',
   '--pattern-text-4xl': '3rem',
-
   '--pattern-scale': '1',
   '--pattern-stroke-xs': '0.2px',
   '--pattern-stroke-sm': '0.4px',
