--- sites/shared/themes/index.js
@@ -1,16 +1,0 @@
-// This is CJS because Tailwind does not (yet) support ESM
-const light = require('./light')
-const dark = require('./dark')
-const hax0r = require('./hax0r')
-const lgbtq = require('./lgbtq')
-const pastel = require('./pastel')
-const aqua = require('./aqua')
-
-module.exports = {
-  light,
-  dark,
-  aqua,
-  hax0r,
-  lgbtq,
-  pastel,
-}
