--- sites/shared/themes/runtime.mjs
@@ -1,22 +1,0 @@
-export const themes = {
-  light: {
-    base: '#fafafa',
-    dark: '#171717',
-  },
-  dark: {
-    base: '#171717',
-    dark: '#000000',
-  },
-  hax0r: {
-    base: '#002908',
-    dark: '#001a06',
-  },
-  lgbtq: {
-    base: '#fafafa',
-    dark: '#171717',
-  },
-  trans: {
-    base: '#fafafa',
-    dark: '#171717',
-  },
-}
